refactor: Extract SASL mechanism registry into SaslMechanismCatalog
Moves the mechanism registry, its configuration property listener, and the SASL provider registration out of SASLAuthentication. The originals remain as deprecated delegates. No behaviour change.

Moves the registry tests out of SASLIntegrationTest into SaslMechanismCatalogTest, which rebuilds the registry per test rather than leaking state between them, and adds coverage for the distinction between enabled, implemented and supported mechanisms.
